ENH: Added a new user method
This method will allow a user to query Netbox with a csv and return each
device name and if they are in Netbox or not.
